requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
boto3==1.17.58
botocore==1.20.112
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
Django==3.2.6
django-cors-headers==3.8.0
django-dotenv==1.4.2
django-environ==0.6.0
django-storages==1.11.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
idna==3.3
jmespath==0.10.0
mysqlclient==2.0.3
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.3.2
pyasn1==0.4.8
pycparser==2.21
PyPDF2==1.26.0
python-barcode==0.13.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
razorpay==1.2.0
reportlab==3.6.3
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
